Split task service to manager and ttrpc service
