+++ cmd/containerd-shim-nerdbox-v1/main.go
@@ -29,6 +29,7 @@ import (
 	_ "github.com/containerd/nerdbox/plugins/shim/streaming"
 	_ "github.com/containerd/nerdbox/plugins/shim/task"
 	_ "github.com/containerd/nerdbox/plugins/shim/transfer"
+	_ "github.com/containerd/nerdbox/plugins/task"
 	_ "github.com/containerd/nerdbox/plugins/vm/libkrun"
 )
