+++ plugins/shim/task/plugin.go
@@ -20,61 +20,14 @@ import (
 	"fmt"
 
 	taskAPI "github.com/containerd/containerd/api/runtime/task/v3"
-	"github.com/containerd/containerd/v2/pkg/shim"
-	"github.com/containerd/containerd/v2/pkg/shutdown"
-	cplugins "github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/plugin"
 	"github.com/containerd/plugin/registry"
 	"github.com/containerd/ttrpc"
 
-	intsandbox "github.com/containerd/nerdbox/internal/shim/sandbox"
-	"github.com/containerd/nerdbox/internal/shim/task"
 	"github.com/containerd/nerdbox/plugins"
 )
 
 func init() {
-	registry.Register(&plugin.Registration{
-		Type: plugins.TaskPlugin,
-		ID:   "manager",
-		Requires: []plugin.Type{
-			cplugins.EventPlugin,
-			cplugins.InternalPlugin,
-			plugins.SandboxPlugin,
-		},
-		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			pp, err := ic.GetByID(cplugins.EventPlugin, "publisher")
-			if err != nil {
-				return nil, err
-			}
-			ss, err := ic.GetByID(cplugins.InternalPlugin, "shutdown")
-			if err != nil {
-				return nil, err
-			}
-			sbRaw, err := ic.GetSingle(plugins.SandboxPlugin)
-			if err != nil {
-				return nil, err
-			}
-
-			svc, ok := sbRaw.(*intsandbox.SandboxService)
-			if !ok {
-				return nil, fmt.Errorf("unexpected SandboxPlugin implementation %T", sbRaw)
-			}
-
-			// Determine debug flag from shim opts stored in context.
-			debug := false
-			if opts, ok := ic.Context.Value(shim.OptsKey{}).(shim.Opts); ok {
-				debug = opts.Debug
-			}
-
-			// Wire the bundle-derived VM start options callback into the
-			// SandboxService so that StartSandbox can boot the VM with the
-			// correct resources and networking without importing the task package.
-			svc.RegisterStartOptions(task.SandboxStartOptions(debug))
-
-			return task.NewTaskService(ic.Context, svc, pp.(shim.Publisher), ss.(shutdown.Service))
-		},
-	})
-
 	registry.Register(&plugin.Registration{
 		Type: plugins.TTRPCPlugin,
 		ID:   "task",
